Changes to api_reference.md subheadings (and ToC)
User research feedback suggested that when looking at the ToC, users felt the headings such as 'View a specific record' and 'View all entires' weren't technical enough. The changes in this commit change the subheadings (and therefore the ToC) to more technical names, to aid searching. This will be fed into the next round of UR in week 9 of Q3 to ascertain whether in this new form, this is easier for users to follow.