Change "Rate limits" to "No rate limits"
It's a bit unorthodox to do this, but since it is so clear that we have no rate limits enforced, I don't see why we shouldn't just make this very simple and straightforward change. This isn't necessarily based on research but it's very minor in real terms, and is highly unlikely to impact the user journey. 

This PR also removes the "contact us" element of `rate_limits.md`, which should not be necessary since there is a separate section for that.