Example change
The second example in the 'See what data is in each entry' section was incorrect for this PR, I realised.

On reflection, do we need it? If we go with this version of the content (i.e. what's in this 'spinoff' PR) for the 'Getting updates' section then I would need to truncate the response for the above call on the `government-organisation` request, in order to show an entry for which there was an example call on an item hash using `GET /items/{item-hash}`.

I actually think linking to the relation section as I suggest in this PR is better. Do you have any thoughts @arnau @michaelabenyohai?