Remove 'therubyracer' from Gemfile
This was added because of https://github.com/alphagov/registers-tech-docs/pull/47#pullrequestreview-111624149 but it isn't necessary any more to build, and has actually been linked to recent hangs/crashes of the tech docs tool, so I'd like to remove it. It shouldn't have any impact, but if it does, it's an easily identifiable fix.