Document new /blobs endpoint
I've also restructured this page to work better with the navigation
sidebar.

The intention is that when we release this new API version,
we will rename this thing from "Upgrading from v1 to next" to
"Upgrading from v1 to v2", and add sections as needed when
creating new versions.

At some point we will also need to create a page for `GET /blobs`
but in order to do this we need a way of structuring content by
version number. How we do this depends on whether we accept
https://github.com/openregister/registers-rfcs/pull/43 and
just autogenerate the reference documentation.